improve airbyte ci gradle
